chore(web): move correction-search execution timer to its own file
